[domain] Move SyncTotals into module
