+++ core/src/main/scala/net/kemitix/thorp/core/SyncPlan.scala
@@ -1,4 +1,6 @@
 package net.kemitix.thorp.core
 
+import net.kemitix.thorp.domain.SyncTotals
+
 case class SyncPlan(actions: Stream[Action] = Stream(),
                     syncTotals: SyncTotals = SyncTotals())
