+++ domain/src/main/scala/net/kemitix/thorp/domain/SyncTotals.scala
@@ -1,4 +1,4 @@
-package net.kemitix.thorp.core
+package net.kemitix.thorp.domain
 
 case class SyncTotals(count: Long = 0L,
                       totalSizeBytes: Long = 0L)
